Initial commit
The completed .Rmd file and a README for the assignment. Unfortunately,
knitting the HTML file is taking a lot longer than I had hoped, and it
will probably not be ready in time for the deadline for the assignment.
I will get the HTML file posted as soon as I can, but I will completely
understand if you award no points for this submission.